Remove `getType` in favour of a static method which will ensure StorageType is always local.

+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/type/TextType.java
@@ -37,7 +37,7 @@ public final class TextType implements StorageType<String> {
   }
 
   @Override
-  public long size() {
+  public int size() {
     return maxLength;
   }
